requirements.txt
kopf==1.35.6
kubernetes==24.2.0
loguru==0.6.0
PyYAML==6.0
jinja2==3.1.2
bcrypt==4.0.1
pykube-ng[gcp]==22.9.0
requests
